+++ tests/smoke/blt_hip_gtest_smoke.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2024, Lawrence Livermore National Security, LLC and
+// Copyright (c) 2017-2025, Lawrence Livermore National Security, LLC and
 // other BLT Project Developers. See the top-level LICENSE file for details
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
@@ -8,14 +8,11 @@
 // file: blt_hip_smoke.cpp
 //
 //-----------------------------------------------------------------------------
-
-#include <iostream>
-#include <stdio.h>
 #include "gtest/gtest.h"
 #include "hip/hip_runtime.h"
 
 __device__ const char STR[] = "HELLO WORLD!";
-const char STR_LENGTH = 12;
+const int STR_LENGTH = 12;
 
 __global__ void hello()
 {
@@ -27,12 +24,13 @@ __global__ void hello()
 //------------------------------------------------------------------------------
 TEST(blt_hip_gtest_smoke,basic_assert_example)
 {
+  hipError_t rc;
   int num_threads = STR_LENGTH;
   int num_blocks = 1;
   hipLaunchKernelGGL((hello), dim3(num_blocks), dim3(num_threads),0,0);
-  if(hipSuccess != hipDeviceSynchronize())
-  {
-    std::cout << "ERROR: hipDeviceSynchronize failed!" << std::endl;
-  }
-  EXPECT_TRUE( true );
+  rc = hipGetLastError();
+  EXPECT_EQ(rc,hipSuccess) << "[HIP ERROR]: " << hipGetErrorString(rc) << "\n";
+
+  rc = hipDeviceSynchronize();
+  EXPECT_EQ(rc,hipSuccess) << "[HIP ERROR]: " << hipGetErrorString(rc) << "\n";
 }