+++ tests/smoke/blt_hip_runtime_c_smoke.c
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2024, Lawrence Livermore National Security, LLC and
+// Copyright (c) 2017-2025, Lawrence Livermore National Security, LLC and
 // other BLT Project Developers. See the top-level LICENSE file for details
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
@@ -15,13 +15,26 @@
 
 int main()
 {
-  int nDevices;
+  hipError_t rc = hipSuccess;     
+  int nDevices = 0;
 
-  hipGetDeviceCount(&nDevices);
+  rc = hipGetDeviceCount(&nDevices);
+  if (rc != hipSuccess)
+  {
+    fprintf(stderr, "[HIP ERROR]: %s", hipGetErrorString(rc));
+    return -1;
+  }
+
   for (int i = 0; i < nDevices; i++)
   {
     hipDeviceProp_t prop;
-    hipGetDeviceProperties(&prop, i);
+    rc = hipGetDeviceProperties(&prop, i);
+    if (rc != hipSuccess)
+    {
+      fprintf(stderr, "[HIP ERROR]: %s", hipGetErrorString(rc));
+      return -1;
+    }
+
     printf("Device Number: %d\n", i);
     printf("  Device name: %s\n", prop.name);
     printf("  Memory Clock Rate (KHz): %d\n",